reuse ByteBufRefCounted
